Getting closer to a usable app.
BlogEngine is working, the settings to add new accounts is getting
pretty close. I have to figure out why CoreData throws an exception when
the new account comes, but otherwise things are working. Also did some
work with the blog post actions, so now you can double click to view an
article in the browser.